Use global compile options
The previous iteration used target_compile_options, but this meant that
only the main file was compiled with these flags.  This changes the
settings for C and C++ so that the flags are applied globally instead,
so that for projects that are composed of multiple source files, all
source files will use the same settings.

Additionally, the default C++ standard used was bumped from 17 to 20.

Signed-off-by: Ed Beroset <beroset@ieee.org>